transfer all the functions from minian_utilites to minian_main
